+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Warning: Before deploying in production a bot that can generate heavy traffic, p
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
 ## [2.22.1] - 2023-05-XX
--   Add  "userId" and "description" when creating AlertDevice.
+-   Add "userId" and "description" when creating AlertDevice.
+-   Add private API for PBX : `AdminService::createPbxPhoneNumber`, `AdminService::deletePbxPhoneNumber`, `AdminService::getPbxPhoneNumber`, `AdminService::getAllPbxPhoneNumbers`, `AdminService::updatepbxPhoneNumber`
 
 ## [2.22.1] - 2023-05-11
 -   Fix `RpcoverxmppEventHandler::_onIqGetSetQueryReceived`.